requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm==1.0.24
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
google-genai>=1.0.0
grad-cam==1.5.4
Pillow>=11.0.0
numpy>=1.24.0
opencv-python-headless==4.10.0.84
python-dotenv==1.0.1
fpdf2==2.8.2
streamlit>=1.40.0
streamlit-paste-button==0.1.2
matplotlib>=3.7.0
huggingface-hub>=0.20.0
requests>=2.31.0
